# api/internal/delivery/http/api.gen.go
//go:build go1.22

// Package httpdelivery provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package httpdelivery

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for CreateSecretRequestVersion.
const (
	N1 CreateSecretRequestVersion = 1
)

// Defines values for HealthResponseStatus.
const (
	Ok HealthResponseStatus = "ok"
)

// CreateSecretRequest defines model for CreateSecretRequest.
type CreateSecretRequest struct {
	// Ciphertext Raw, unpadded base64url-encoded ciphertext.
	Ciphertext string    `json:"ciphertext"`
	ExpiresAt  time.Time `json:"expiresAt"`

	// Nonce Raw, unpadded base64url-encoded 12-byte nonce.
	Nonce string `json:"nonce"`

	// SecretId Base64url-encoded identifier containing at least 16 bytes.
	SecretId string                     `json:"secretId"`
	Version  CreateSecretRequestVersion `json:"version"`
}

// CreateSecretRequestVersion defines model for CreateSecretRequest.Version.
type CreateSecretRequestVersion int

// CreateSecretResponse defines model for CreateSecretResponse.
type CreateSecretResponse struct {
	SecretId string `json:"secretId"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string `json:"error"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status HealthResponseStatus `json:"status"`
}

// HealthResponseStatus defines model for HealthResponse.Status.
type HealthResponseStatus string

// SecretResponse defines model for SecretResponse.
type SecretResponse struct {
	Ciphertext string    `json:"ciphertext"`
	ExpiresAt  time.Time `json:"expiresAt"`
	Nonce      string    `json:"nonce"`
	SecretId   string    `json:"secretId"`
	Version    int       `json:"version"`
}

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// HealthOK defines model for HealthOK.
type HealthOK = HealthResponse

// InternalError defines model for InternalError.
type InternalError = ErrorResponse

// SecretUnavailable defines model for SecretUnavailable.
type SecretUnavailable = ErrorResponse

// ValidationError defines model for ValidationError.
type ValidationError = ErrorResponse

// CreateSecretJSONRequestBody defines body for CreateSecret for application/json ContentType.
type CreateSecretJSONRequestBody = CreateSecretRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Check whether the process is alive
	// (GET /api/v1/health/live)
	LiveHealth(w http.ResponseWriter, r *http.Request)
	// Check whether the process is ready
	// (GET /api/v1/health/ready)
	ReadyHealth(w http.ResponseWriter, r *http.Request)
	// Create a one-time secret
	// (POST /api/v1/secrets)
	CreateSecret(w http.ResponseWriter, r *http.Request)
	// Claim a one-time secret
	// (POST /api/v1/secrets/{secretID}/claim)
	ClaimSecret(w http.ResponseWriter, r *http.Request, secretID string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// LiveHealth operation middleware
func (siw *ServerInterfaceWrapper) LiveHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.LiveHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ReadyHealth operation middleware
func (siw *ServerInterfaceWrapper) ReadyHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ReadyHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateSecret operation middleware
func (siw *ServerInterfaceWrapper) CreateSecret(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateSecret(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ClaimSecret operation middleware
func (siw *ServerInterfaceWrapper) ClaimSecret(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "secretID" -------------
	var secretID string

	err = runtime.BindStyledParameterWithOptions("simple", "secretID", r.PathValue("secretID"), &secretID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "secretID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ClaimSecret(w, r, secretID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/api/v1/health/live", wrapper.LiveHealth)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/health/ready", wrapper.ReadyHealth)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/secrets", wrapper.CreateSecret)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/secrets/{secretID}/claim", wrapper.ClaimSecret)

	return m
}

type BadRequestJSONResponse ErrorResponse

type HealthOKJSONResponse HealthResponse

type InternalErrorJSONResponse ErrorResponse

type SecretUnavailableJSONResponse ErrorResponse

type ValidationErrorJSONResponse ErrorResponse

type LiveHealthRequestObject struct {
}

type LiveHealthResponseObject interface {
	VisitLiveHealthResponse(w http.ResponseWriter) error
}

type LiveHealth200JSONResponse struct{ HealthOKJSONResponse }

func (response LiveHealth200JSONResponse) VisitLiveHealthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ReadyHealthRequestObject struct {
}

type ReadyHealthResponseObject interface {
	VisitReadyHealthResponse(w http.ResponseWriter) error
}

type ReadyHealth200JSONResponse struct{ HealthOKJSONResponse }

func (response ReadyHealth200JSONResponse) VisitReadyHealthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type CreateSecretRequestObject struct {
	Body *CreateSecretJSONRequestBody
}

type CreateSecretResponseObject interface {
	VisitCreateSecretResponse(w http.ResponseWriter) error
}

type CreateSecret201JSONResponse CreateSecretResponse

func (response CreateSecret201JSONResponse) VisitCreateSecretResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateSecret400JSONResponse struct{ BadRequestJSONResponse }

func (response CreateSecret400JSONResponse) VisitCreateSecretResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateSecret422JSONResponse struct{ ValidationErrorJSONResponse }

func (response CreateSecret422JSONResponse) VisitCreateSecretResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)

	return json.NewEncoder(w).Encode(response)
}

type CreateSecret500JSONResponse struct{ InternalErrorJSONResponse }

func (response CreateSecret500JSONResponse) VisitCreateSecretResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type ClaimSecretRequestObject struct {
	SecretID string `json:"secretID"`
}

type ClaimSecretResponseObject interface {
	VisitClaimSecretResponse(w http.ResponseWriter) error
}

type ClaimSecret200JSONResponse SecretResponse

func (response ClaimSecret200JSONResponse) VisitClaimSecretResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ClaimSecret404JSONResponse struct{ SecretUnavailableJSONResponse }

func (response ClaimSecret404JSONResponse) VisitClaimSecretResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type ClaimSecret500JSONResponse struct{ InternalErrorJSONResponse }

func (response ClaimSecret500JSONResponse) VisitClaimSecretResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Check whether the process is alive
	// (GET /api/v1/health/live)
	LiveHealth(ctx context.Context, request LiveHealthRequestObject) (LiveHealthResponseObject, error)
	// Check whether the process is ready
	// (GET /api/v1/health/ready)
	ReadyHealth(ctx context.Context, request ReadyHealthRequestObject) (ReadyHealthResponseObject, error)
	// Create a one-time secret
	// (POST /api/v1/secrets)
	CreateSecret(ctx context.Context, request CreateSecretRequestObject) (CreateSecretResponseObject, error)
	// Claim a one-time secret
	// (POST /api/v1/secrets/{secretID}/claim)
	ClaimSecret(ctx context.Context, request ClaimSecretRequestObject) (ClaimSecretResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// LiveHealth operation middleware
func (sh *strictHandler) LiveHealth(w http.ResponseWriter, r *http.Request) {
	var request LiveHealthRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.LiveHealth(ctx, request.(LiveHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "LiveHealth")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(LiveHealthResponseObject); ok {
		if err := validResponse.VisitLiveHealthResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ReadyHealth operation middleware
func (sh *strictHandler) ReadyHealth(w http.ResponseWriter, r *http.Request) {
	var request ReadyHealthRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ReadyHealth(ctx, request.(ReadyHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ReadyHealth")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ReadyHealthResponseObject); ok {
		if err := validResponse.VisitReadyHealthResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateSecret operation middleware
func (sh *strictHandler) CreateSecret(w http.ResponseWriter, r *http.Request) {
	var request CreateSecretRequestObject

	var body CreateSecretJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateSecret(ctx, request.(CreateSecretRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateSecret")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateSecretResponseObject); ok {
		if err := validResponse.VisitCreateSecretResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ClaimSecret operation middleware
func (sh *strictHandler) ClaimSecret(w http.ResponseWriter, r *http.Request, secretID string) {
	var request ClaimSecretRequestObject

	request.SecretID = secretID

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ClaimSecret(ctx, request.(ClaimSecretRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ClaimSecret")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ClaimSecretResponseObject); ok {
		if err := validResponse.VisitClaimSecretResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}